requirements.txt
# Core dependencies for Interactive CSV Analyzer
openai>=1.40.0
e2b-code-interpreter>=1.0.0
python-dotenv>=1.0.0

# Data analysis and visualization
pandas>=2.0.0
matplotlib>=3.7.0
seaborn>=0.12.0
numpy>=1.24.0

# Optional enhancements
jupyter>=1.0.0
ipython>=8.0.0
